hospitalapi/requirements.txt
anyio==3.6.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
paynow==1.0.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
